+++ simulator/support/esp32-hal.h
@@ -7,7 +7,7 @@
 #define PI 3.1415926535897932384626433832795
 
 int millis();
-int esp_timer_get_time();
+int64_t esp_timer_get_time();
 int constrain(int val, int min, int max);
 int random8(int lim = 255);
 int random16(int lim = 255*255);
